Add bandwidth bound on k-incoherence
By [Linear Algebra and its Applications 716, 32–59 (2025)], we know that the minimum bandwidth of a matrix up to symmetric permutation is an upper bound on its factor width. (Recall that a matrix is said to be k-incoherent if its factor width is at most k.) We thus implement has_bandwidth_k_ordering.m in helpers/ and add this as a dependency for IskIncoherent.m, which calls the algorithm to check whether bandwidth <= k (applying a recognition algorithm, not a minimization one) and returning true if so.

NOTE: There are also a few bugs in the IskIncoherent logic and documentation unrelated to bandwidth; we shall address this in a follow-up PR to maintain separation of concerns.