Fix bugs and docs in IskIncoherent.m
This PR fixes the bug in IskIncoherent arising from nesting the CVX optimization problem (which created scoping issues for the variable A). It also fixes the docs to align better with the current implementation--previously, -1 was sometimes returned when the function could not produce a definitive answer, but the addition of SDP ensures that we always find an answer. (References are also fixed/updated.)

Finally, CHANGELOG.md is updated to reflect both this and the bandwidth change from #34 (which forgot to update the changelog).